# faster iteration over subarrays (views)