--- a/base/subarray.jl
+++ b/base/subarray.jl
@@ -433,28 +433,32 @@ find_extended_inds(::ScalarIndex, I...) = (@inline; find_extended_inds(I...))
 find_extended_inds(i1, I...) = (@inline; (i1, find_extended_inds(I...)...))
 find_extended_inds() = ()
 
-function isvalid_index(p, i, l)
-    if firstindex(p) == 1
-        (i-1)%UInt < l%UInt <= lastindex(p)%UInt
+function is_valid_ind(p, i, l, stride1)
+    fi = firstindex(p)
+    li = lastindex(p)
+    if stride1 < 0
+        fi <= l <= i <= li
+    elseif fi != 1
+        fi <= i <= l <= li
     else
-        firstindex(p) <= i <= l <= lastindex(p)
+        (i-1) % UInt < l % UInt <= li % UInt
     end
 end
 
-function Base.iterate(v::Base.FastContiguousSubArray,
+function iterate(v::FastContiguousSubArray,
         state = v.offset1 .+ (firstindex(v), lastindex(v)))
     p = parent(v)
     i, l = state
-    isvalid_index(p, i, l) ? (@inbounds p[i], (i+1, l)) : nothing
+    is_valid_ind(p, i, l, 1) ? (@inbounds p[i], (i+1, l)) : nothing
 end
 
 stride1(v::SubArray{T,N,P}) where {T,N,P} = P <: StridedArray ? stride(v, 1) : v.stride1
 
-function Base.iterate(v::Base.FastSubArray,
+function iterate(v::FastSubArray,
         state = v.offset1 .+ stride1(v) .* (firstindex(v), lastindex(v)))
     p = parent(v)
     i, l = state
-    isvalid_index(p, i, l) ? (@inbounds p[i], (i+stride1(v), l)) : nothing
+    is_valid_ind(p, i, l, stride1(v)) ? (@inbounds p[i], (i+stride1(v), l)) : nothing
 end
 
 function unsafe_convert(::Type{Ptr{T}}, V::SubArray{T,N,P,<:Tuple{Vararg{RangeIndex}}}) where {T,N,P}