@N5N3 2023-02-19
We just need `@inbounds` to make SIMD works in this case.
LLVM is clever enough to optimize the index transformation away.
But it's hard for it to prove the access is inbounds as SubArray store's it's own indices.