@Seelengrab 2023-02-19
Does SIMDing in `foldl` preserve the foldl order of operations? I'm not sure we want to use `@fastmath` by default, since it allows reassociations and can lead to [very surprising](https://discourse.julialang.org/t/array-ordering-and-naive-summation/1929) differences in results.