@matthias314 2023-02-19
Note that I'm using `foldl` just as an example for an iteration.

> We just need `@inbounds` to make SIMD works in this case.

@N5N3 Are you sure? On master the timings for
```
f(v) = @inbounds foldl(+, v; init = zero(eltype(v)))
```
are the same as those without `@inbounds`.

> Also, this introduces inbounds into a view that may wrap any AbstractArray, including one with a mis-implementation of length. I don't think this is safe.

@jakobnissen  It's only unsafe for a view into some `T <: AbstractArray` if `getindex` is defined for `T` with `@propagate_inbounds`. I would argue that one should only use `@propagate_inbounds` if one is sure that the implementation is correct.

> Can you check if LLVM can omit the boundscheck automatically?

If I drop the `@inbounds` in `iterate`, I get the following timings:
```
n = 2^16

v = rand(Int8, n);
w = view(v, :);

julia> @btime f($v)
  568.296 ns (0 allocations: 0 bytes)
julia> @btime f($w)
  21.947 μs (0 allocations: 0 bytes)

s = String(rand('a':'z', n));
u = codeunits(s);
v = view(u, :);

julia> @btime f($u)
  562.832 ns (0 allocations: 0 bytes)
julia> @btime f($v)
  662.285 ns (0 allocations: 0 bytes)
```
Note that for `CodeUnits` they are very similar, but not for `Vector`. For `Vector` I can still see
```
movabs	rax, offset ijl_bounds_error_ints
```
in `@code_native f(w)`.

> Could this be generalized to views with integer StepRange indices, such that it also works for `@view a[1:2:5]` (i.e FastSubArray not necessary contiguous)?

One problem I see is that the step value would not be known at compile time. More importantly, are there SIMD instructions that can handle non-contiguous data?