@jakobnissen 2023-02-19
Good point regarding safety, and indeed it looks like LLVM can't omit boundscheck.

It's true that the step range would not be known at compile time (except for contiguous arrays where the step size is zero). But you would still have the benefit of inbounds. And it means you can potentially use a single implementation for contiguous and non-contiguous views with no runtime cost.