@matthias314 2023-02-19
> you can potentially use a single implementation for contiguous and non-contiguous views with no runtime cost

For that one needs to change `iterate` to something linke
```
function Base.iterate(v::Base.FastSubArray,
        state = (v.offset1+v.stride1*firstindex(v), v.offset1+v.stride1*lastindex(v)))
    i, l = state
    @inbounds i-1 < l ? (v.parent[i], (i+v.stride1, l)) : nothing
end
```
Then there is much smaller performance gain for `FastContiguousSubArray`,
```
n = 2^16; v = rand(Int8, n); w = view(v, :);

julia> @btime f($v)
  565.796 ns (0 allocations: 0 bytes)
julia> @btime f($w)
  34.617 μs (0 allocations: 0 bytes)   # master
  21.928 μs (0 allocations: 0 bytes)   # with iterate as above
```
For `FastSubArray`, one gets as fast as `mapfoldl` in this case:
```
r = 2:2:length(v); w = view(v, r);   # v as above

julia> @btime mapfoldl(i -> @inbounds($v[i]), +, $r; init = zero(eltype($v)))
  10.963 μs (0 allocations: 0 bytes)
julia> @btime f($w)
  190.195 μs (0 allocations: 0 bytes)   # master
  10.980 μs (0 allocations: 0 bytes)    # with iterate as above
```
However, one could achieve the same by adding `@inbounds` to the `iterate` method for `AbstractArray`, which is the one currently used for `SubArray`. (I'm omitting code and benchmark here.) That idea was discussed in #40397 and not implemented in the end. In light of the above performance difference, one might have wished for a different outcome.