@N5N3 2023-02-20
> are the same as those without @inbounds.

I'm sure. Non of `foldl` and `iterate` propagates `@inbounds` thus `@inbounds foldl` won't omit the bounds check.
To make this work, you need:
```julia
struct SafeArray{T,N,P<:AbstractArray{T,N}} <: AbstractArray{T,N}
    parent::P
end
@inline Base.size(x::SafeArray) = Base.size(x.parent)
@inline Base.getindex(x::SafeArray, is...) = @inbounds x.parent[is...]
```
Then you will get
```julia
julia> @btime f($a)
  244.975 ns (0 allocations: 0 bytes)
157838443398623711

julia> @btime f($v)
  1.480 μs (0 allocations: 0 bytes)
157838443398623711

julia> @btime f(SafeArray($v))
  249.086 ns (0 allocations: 0 bytes)
157838443398623711
```