@jakobnissen 2023-02-20
@matthias314 When I use the function `stride(v, 1)` instead of accessing `v.stride` as a field, for me it constant folds, and I get maximal performance from the contiguous arrays, while also being as fast for non-contiguous arrays. Here is the function:
```
function Base.iterate(v::Base.FastSubArray,
    state = (v.offset1+stride(v,1)*firstindex(v), v.offset1+stride(v,1)*lastindex(v)))
    i, l = state
    @inbounds i-1 < l ? (v.parent[i], (i+stride(v,1), l)) : nothing
end
```
And the benchmarks:
```
f(v) = foldl(+, v; init = zero(eltype(v)))
a = rand(Int8, 2^16);
v = view(a, :);

julia> @btime f($a);
  485.785 ns (0 allocations: 0 bytes)

julia> @btime f($v);
  489.887 ns (0 allocations: 0 bytes)
```