@N5N3 2023-02-20
`stride1` for `FastSubArray` is specific for the parent while `stride(_, 1)` is specific for the memory layout.
They can be different and `strides` might be undefined for `FastSubArray` if its parent has no `strides` defined.

BTW, I don't think this PR could be finished with out the added `@inbounds`.
As #40397 was closed with "this is too unsafe", I think the correct direction is further restrict the function sig to e.g. `StridedArray`