@matthias314 2023-02-20
Thanks a lot for your suggestions! Some of them seem to require end-users to rewrite their code (which was not what I had in mind), but they are all very instructive.

I agree that nothing can be done without using `@inbounds`, and that may not fly in general. However, I'm reluctant to give up completely. We are talking about potential speed-up factors of 60 (`FastContiguousSubArray`) or 17 (`FastSubArray` and I believe `SubArray` in general). That seems to much to let go of it. At least for views into `Array` and maybe other standard types one should do something. What do you think?