@jakobnissen 2023-02-20
I agree 100%, this is a good contribution and will really help. If you restrict the type signature to only views of `Array` with either an `AbstractUnitRange` or `AbstractStepRange`, then most real-life cases will be covered.