@matthias314 2023-02-21
Here is a new version. It's now safe (that is, with bound checks), but still as fast as iterating over the parent array in many cases: For a `FastContiguousSubArray` this holds if it is 1-based. For any other `FastSubArray` `a` it holds if additionally `stride(a,1)` can be determined at compile-time. Here are some benchmarks:
```
n = 2^16; m = 2^8

v = rand(Int8, n)
u = codeunits(String(rand('a':'z', n)))
ov = OffsetVector(v, -n:-1)
oa = OffsetMatrix(rand(Int8, m, m), -m:-1, -m:-1)

f(v) = foldl(+, v; init = zero(eltype(v)))
```
In each case, `f` is applied to the array.

| parent array | without view | view (master) | view (this PR) |
|---|---|---|---|
| `view(v, :)` | 573.613 ns | 34.615 μs | 571.708 ns |
| `@view v[2:2:end]` | 10.962 μs | 190.205 μs | 10.980 μs |
| `view(u,:)` | 577.000 ns | 27.047 μs | 567.232 ns |
| `view(ov,:)` | 581.723 ns | 40.696 μs | 18.320 μs |
| `view(oa,:,:)` | 578.016 ns | 34.615 μs | 588.478 ns |

The reason that `OffsetVector` is so much slower than a higher-dimensional `OffsetArray` is that in the higher-dimensional case the linear indexing is always 1-based while for a vector it is the given indexing.